feat: auto-create table when missing
Add option `influxdb.auto_create_table` that when set will automatically
create a table by the name given by the measurement. The definition of
the table will consist of a `_time` column, a `_tags` column, and a
`_fields` column.

This will allow users to quickly collect measurements and then later
decide on the table definitions.